apps/api/pyproject.toml
[project]
name = "fuse-api"
version = "0.1.0"
description = "Fuse Platform API"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "alembic",
    "asyncpg",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "redis",
    "celery",
    "cryptography",
    "httpx",
]

[tool.uv]
managed = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
